bls12_381 0.8.0 v6 patch

+++ src/fp.rs
@@ -489,8 +489,8 @@ impl Fp {
     pub fn add_inp(&mut self, rhs: &Fp) {
         unsafe {
             syscall_bls12381_fp_addmod(
-                self.0.as_mut_ptr() as *mut u32,
-                rhs.0.as_ptr() as *const u32,
+                self.0.as_mut_ptr() as *mut u64,
+                rhs.0.as_ptr() as *const u64,
             );
         }
     }
@@ -516,7 +516,7 @@ impl Fp {
             if #[cfg(target_os = "zkvm")] {
                 let mut out = self.clone();
                 unsafe {
-                    syscall_bls12381_fp_addmod(out.0.as_mut_ptr() as *mut u32, rhs.0.as_ptr() as *const u32);
+                    syscall_bls12381_fp_addmod(out.0.as_mut_ptr() as *mut u64, rhs.0.as_ptr() as *const u64);
                 }
                 out
             } else {
@@ -556,7 +556,7 @@ impl Fp {
             if #[cfg(target_os = "zkvm")] {
                 let mut out = Fp::zero();
                 unsafe {
-                    syscall_bls12381_fp_submod(out.0.as_mut_ptr() as *mut u32, self.0.as_ptr() as *const u32);
+                    syscall_bls12381_fp_submod(out.0.as_mut_ptr() as *mut u64, self.0.as_ptr() as *const u64);
                 }
                 out
             } else {
@@ -570,8 +570,8 @@ impl Fp {
     pub fn sub_inp(&mut self, rhs: &Fp) {
         unsafe {
             syscall_bls12381_fp_submod(
-                self.0.as_mut_ptr() as *mut u32,
-                rhs.0.as_ptr() as *const u32,
+                self.0.as_mut_ptr() as *mut u64,
+                rhs.0.as_ptr() as *const u64,
             );
         }
     }
@@ -582,7 +582,7 @@ impl Fp {
             if #[cfg(target_os = "zkvm")] {
                 let mut out = self.clone();
                 unsafe {
-                    syscall_bls12381_fp_submod(out.0.as_mut_ptr() as *mut u32, rhs.0.as_ptr() as *const u32);
+                    syscall_bls12381_fp_submod(out.0.as_mut_ptr() as *mut u64, rhs.0.as_ptr() as *const u64);
                 }
                 out
             } else {
@@ -755,8 +755,8 @@ impl Fp {
     pub fn mul_inp(&mut self, rhs: &Fp) {
         unsafe {
             syscall_bls12381_fp_mulmod(
-                self.0.as_mut_ptr() as *mut u32,
-                rhs.0.as_ptr() as *const u32,
+                self.0.as_mut_ptr() as *mut u64,
+                rhs.0.as_ptr() as *const u64,
             );
         }
         self.mul_r_inv_internal();
@@ -816,7 +816,7 @@ impl Fp {
             if #[cfg(target_os = "zkvm")] {
                 let mut out = self.clone();
                 unsafe {
-                    syscall_bls12381_fp_mulmod(out.0.as_mut_ptr() as *mut u32, rhs.0.as_ptr() as *const u32);
+                    syscall_bls12381_fp_mulmod(out.0.as_mut_ptr() as *mut u64, rhs.0.as_ptr() as *const u64);
                 }
                 out.mul_r_inv_internal();
                 out
@@ -834,8 +834,8 @@ impl Fp {
     pub(crate) fn mul_r_inv_internal(&mut self) {
         unsafe {
             syscall_bls12381_fp_mulmod(
-                self.0.as_mut_ptr() as *mut u32,
-                R_INV.0.as_ptr() as *const u32,
+                self.0.as_mut_ptr() as *mut u64,
+                R_INV.0.as_ptr() as *const u64,
             );
         }
     }
@@ -847,7 +847,7 @@ impl Fp {
     #[cfg(target_os = "zkvm")]
     pub(crate) fn mul_r_internal(&mut self) {
         unsafe {
-            syscall_bls12381_fp_mulmod(self.0.as_mut_ptr() as *mut u32, R.0.as_ptr() as *const u32);
+            syscall_bls12381_fp_mulmod(self.0.as_mut_ptr() as *mut u64, R.0.as_ptr() as *const u64);
         }
     }
 
@@ -856,8 +856,8 @@ impl Fp {
     pub fn square_inp(&mut self) {
         unsafe {
             syscall_bls12381_fp_mulmod(
-                self.0.as_mut_ptr() as *mut u32,
-                self.0.as_ptr() as *const u32,
+                self.0.as_mut_ptr() as *mut u64,
+                self.0.as_ptr() as *const u64,
             );
         }
         self.mul_r_inv_internal();
@@ -920,7 +920,7 @@ impl Fp {
             if #[cfg(target_os = "zkvm")] {
                 let mut out = self.clone();
                 unsafe {
-                    syscall_bls12381_fp_mulmod(out.0.as_mut_ptr() as *mut u32, self.0.as_ptr() as *const u32);
+                    syscall_bls12381_fp_mulmod(out.0.as_mut_ptr() as *mut u64, self.0.as_ptr() as *const u64);
                 }
                 out.mul_r_inv_internal();
                 out
